move updateUser from step to flow
